refactor: optimize SKIP_LOKI_TESTS usage and restructure integration tests
- Consolidate Loki connection-dependent tests into single file (loki.integration.test.ts)
- Move metrics tests to monitoring directory (no Loki dependency)
- Remove SKIP_LOKI_TESTS dependency from non-Loki tests
- Delete redundant loki-basic.integration.test.ts file
- Enable metrics and Grafana config tests to run in Docker Compose environment

This reduces excessive test skipping from 26 tests to only those that actually
require Loki connectivity.
